requirements.txt
boto3==1.4.4
botocore==1.5.55
click==6.7
coreapi==2.3.0
coreapi-cli==1.0.6
coreschema==0.0.4
Django==1.11.1
django-auth-ldap==1.2.11
django-crispy-forms==1.6.1
django-dashing==0.3.3
django-filter==1.0.2
django-rest-swagger==2.1.2
django-storages==1.5.2
djangorestframework==3.6.3
docutils==0.13.1
itypes==1.1.0
Jinja2==2.9.6
jmespath==0.9.2
Markdown==2.6.8
MarkupSafe==1.0
openapi-codec==1.3.1
psycopg2==2.7.1
pyldap==2.4.35
python-dateutil==2.6.0
pytz==2017.2
requests==2.14.2
s3transfer==0.1.10
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
